// libs/backend/postgres/main/notification/lib/src/domain/enums/notification-error-reason.enum.ts
export enum NotificationErrorReason {
  UserBlocked = 'user_blocked',
  ChatNotFound = 'chat_not_found',
  ChatRestricted = 'chat_restricted',
  BlockedBot = 'blocked_bot',
  BotCantInitiateConversation = 'bot_cant_initiate_converation',
  MarkedAsBlockedBot = 'marked_as_blocked_bot',
  TelegramUserDeactivated = 'telegram_user_deactivated',
  MarkedAsTelegramUserDeactivated = 'marked_as_telegram_user_deactivated',
  UnknownError = 'unknown_error',
  NetworkError = 'network_error',
  RateLimit = 'rate_limit',
  IncorrectTarget = 'incorrect_target',
  NotFoundMessage = 'not_found_message',
  NotFoundMessageStrategy = 'not_found_message_strategy',
}